feat(prb-pcap): add rayon-parallelized packet normalization with fragment fallback
Implements Segment 18 (Phase 2) - Parallel Packet Normalization:
- Extract stateless normalize_stateless() function for thread-safe packet parsing
- Add NormalizeResult enum to distinguish packets from fragments
- Implement NormalizeBatch::run() using rayon par_iter for parallel normalization
- Add process_fragments() for sequential fragment reassembly fallback
- Add static helpers: parse_loopback_static(), parse_sll2_static(), extract_ips(), extract_transport()
- Tests verify parallel matches sequential output, fragment detection, all linktypes
- Fix clippy warnings: remove unnecessary casts, fix needless borrows, add lifetime annotations

All 36 tests pass. Clippy clean with -D warnings.

+++ crates/prb-pcap/src/parallel/normalize.rs
@@ -33,7 +33,7 @@ impl NormalizeBatch {
             .map(|(idx, pkt)| {
                 (
                     idx,
-                    normalize_stateless(pkt.linktype as u32, pkt.timestamp_us, &pkt.data),
+                    normalize_stateless(pkt.linktype, pkt.timestamp_us, &pkt.data),
                 )
             })
             .collect();
@@ -92,7 +92,7 @@ pub fn process_fragments(
 
     for &idx in fragment_indices {
         let pkt = &packets[idx];
-        match normalizer.normalize(pkt.linktype as u32, pkt.timestamp_us, &pkt.data) {
+        match normalizer.normalize(pkt.linktype, pkt.timestamp_us, &pkt.data) {
             Ok(Some(normalized)) => {
                 result.push(OwnedNormalizedPacket::from_borrowed(&normalized));
             }
